fix: fix correct import paths, import fc as type, organize sections into dedicated folder, fix comments and small code mistakes
